finish: final push.

+++ server/src/main/java/com/catalog/controller/admin/AdminController.java
@@ -2,15 +2,15 @@ package com.catalog.controller.admin;
 
 import com.catalog.constant.MessageConstant;
 import com.catalog.dto.AdminLoginDTO;
+import com.catalog.dto.PageDTO;
 import com.catalog.entity.Card;
 import com.catalog.entity.Img;
+import com.catalog.entity.Page;
 import com.catalog.result.Result;
-import com.catalog.service.AdminService;
-import com.catalog.service.CardService;
-import com.catalog.service.ImgService;
-import com.catalog.service.MsgService;
-import com.catalog.utils.MathUtil;
+import com.catalog.service.*;
+import com.catalog.service.Impl.PageServiceImpl;
 import com.catalog.vo.AdminLoginVO;
+import com.catalog.vo.CardListVO;
 import com.catalog.vo.CardVO;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -35,6 +35,8 @@ public class AdminController
     private MsgService msgService;
     @Autowired
     private ImgService imgService;
+    @Autowired
+    private PageService pageService;
 
     @PostMapping("/login")
     @ApiOperation("管理员登入")
@@ -62,10 +64,11 @@ public class AdminController
 
     @GetMapping("/home")
     @ApiOperation("显示主页卡片")
-    public Result<List<CardVO>> showHomeCards()
+    public Result<CardListVO> showHomeCards(@RequestParam(value = "offset", defaultValue = "0") int offset, @RequestParam(value = "limit", defaultValue = "10") int limit)
     {
         List<Card> cards = cardService.getUnAcceptedCard();
         if(cards == null || cards.isEmpty()) return Result.success();
+        int count = cards.size();
         List<CardVO> cardVOs = new ArrayList<>();
         for(Card c : cards)
         {
@@ -82,7 +85,8 @@ public class AdminController
                     .build();
             cardVOs.add(cardVO);
         }
-        return Result.success(cardVOs);
+        cardVOs = cardVOs.subList(Math.min(cardVOs.size(), offset), Math.min(cardVOs.size(), offset+limit));
+        return Result.success(new CardListVO(count, cardVOs));
     }
 
     @DeleteMapping("/reject_card")
@@ -90,8 +94,12 @@ public class AdminController
     public Result<String> deleteRejectCard(@RequestParam(name = "card_id") int cardId)
     {
         log.info("管理员拒绝通过卡片:{}",cardId);
-        adminService.deleteCard(cardId);
+        if(cardService.getCardById(cardId) == null)
+        {
+            return Result.error(MessageConstant.CARD_NOT_EXIST);
+        }
         msgService.sendDeleteMsg(cardId);
+        adminService.deleteCard(cardId);
         return Result.success(MessageConstant.DELETE_SUCCESS);
     }
     @PutMapping("/accept_card")
@@ -99,8 +107,24 @@ public class AdminController
     public Result<String> acceptCard(@RequestParam(name = "card_id") int cardId)
     {
         log.info("管理员允许通过卡片:{}", cardId);
+        Card card = cardService.getCardById(cardId);
+        if(card == null)
+        {
+            return Result.error(MessageConstant.CARD_NOT_EXIST);
+        }
+        if(card.getStatus() == 1)
+        {
+            return Result.error(MessageConstant.CARD_HAVE_ACCEPTED);
+        }
         cardService.acceptCardById(cardId);
         msgService.sendAcceptMsg(cardId);
         return Result.success(MessageConstant.UPDATE_SUCCESS);
+    }
+    @PostMapping("/page")
+    @ApiOperation(("上传官方图鉴"))
+    public Result<String> uploadPage(@RequestBody PageDTO pageDTO)
+    {
+        pageService.uploadPage(pageDTO);
+        return Result.success(MessageConstant.UPLOAD_SUCCESS);
     }
 }
